+++ CHANGELOG.md
@@ -61,7 +61,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
--   **BREAKING** `TreeNode.nodeType` renamed to `TreeNode.nodeTags`.
+-   **BREAKING** `TreeNode.nodeType` renamed to `TreeNode.nodeTags`, in both
+    TypeScript and Scala.
 
 -   We now create a new JS rug for each thread for safety.
     https://github.com/atomist/rug/issues/78